requirements.txt
streamlit
geopandas
pandas
numpy
plotly
openai
python-dotenv
serpapi
langchain-core
langchain
langchain-openai
langgraph
langchain-community
langchain-experimental
typing-extensions
bs4
pandas
matplotlib
seaborn
scikit-learn
numpy
geopandas
shapely
spacy
joblib